refactor: melhorei as classes de serviço

+++ src/main/java/com/book/progress/service/ArchievementService.java
@@ -15,54 +15,39 @@ public class ArchievementService {
     private ArchievementRepository archivementRepository;
 
     //LISTAR TODAS AS CONQUISTAS
-    public List<Archievement> findAllArchivement(){
+    public List<Archievement> findAllArchivement() {
         return archivementRepository.findAll();
     }
 
     //BUSCAR A CONQUISTA POR ID
-    public Optional<Archievement> findIdArchivement(Long id){
+    public Optional<Archievement> findIdArchivement(Long id) {
         return archivementRepository.findById(id);
     }
 
     //SALVAR UMA NOVA CONQUISTA
-    public Archievement saveArquievement(Archievement archievement){
+    public Archievement saveArquievement(Archievement archievement) {
         archievement.setCustom(true);//LOMBOK GERA O SET SEM O "IS"
 
         return archivementRepository.save(archievement);
     }
 
     //ATUALIZAR PONTOS DE UMA CONQUISTA PERSONALIZADA
-    public Optional<Archievement> updateArchievementPoints(Long id, Integer newPoints){
-        Optional<Archievement> optionalArchievement = archivementRepository.findById(id);
-
-        if (optionalArchievement.isPresent()){
-            Archievement archievement = optionalArchievement.get();
-
-            if (archievement.isCustom()){
-                archievement.setPoints(newPoints);
-                archivementRepository.save(archievement);
-
-                return Optional.of(archievement);
-            }
+    public Optional<Archievement> updateArchievementPoints(Long id, Integer newPoints) {
+        Archievement archievement = archivementRepository.findById(id).orElse(null);
+        if (archievement != null && archievement.isCustom()) {
+            archievement.setPoints(newPoints);
+            return Optional.of(archivementRepository.save(archievement));
         }
-
         return Optional.empty();
     }
 
     //DELETAR UMA CONQUISTA PERSONALIZADA
-    public boolean deleteArchievement(Long id){
-        Optional<Archievement> optionalArchievement = archivementRepository.findById(id);
-
-        if (optionalArchievement.isPresent()){
-            Archievement archievement = optionalArchievement.get();
-
-            if (archievement.isCustom()){
-                archivementRepository.delete(archievement);
-
-                return true;
-            }
+    public boolean deleteArchievement(Long id) {
+        Archievement archievement = archivementRepository.findById(id).orElse(null);
+        if (archievement != null && archievement.isCustom()) {
+            archivementRepository.delete(archievement);
+            return true;
         }
-
         return false;
     }
 }
